split off the richtextconverter and opened it up for non-binding conversion, then updated the tooltip to use richtext.
